relocate iframe component
